Add the possibility to convert all column names to lower case.
In MySQL column names are case insensitive.
PostgreSQL folds column names in queries to lower case,
but considers column names with upper case letters in them as different,
requiring the use of double quotes around the column name
and using the proper case when refering to that column.
That causes queries that worked on MySQL to fail on PostgreSQL.
Converting all column names to lower case during migration
avoids this issue.